updated the home page
